- (agent_trap.c): make uptime a long for 64bit architectures.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@3594 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/agent_trap.c
@@ -194,7 +194,7 @@ void send_enterprise_trap_vars (int trap,
     struct variable_list *v2_vars, *last_var=NULL;
     struct snmp_pdu	*template_pdu, *pdu;
     struct timeval	 now;
-    int uptime;
+    long uptime;
     struct sockaddr_in *pduIp;
     struct trap_sink *sink;
